+++ src/Materiel/Planche.java
@@ -1,7 +1,7 @@
 package src.Materiel;
 
 public class Planche extends Commande {
-    public Planche(int id, int nombre, String date, int price, int L, int l){
+    public Planche(int id, int nombre, String date, float price, float L, float l){
         super(id,nombre,date,price,L,l);
     }
 }
